Move : 스타일 폴더 정리
