+++ src/components/board/style/Board.style.jsx
