+++ src/components/common/style/Logo.style.jsx
