+++ src/components/document/Question.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import apiModule from "../../api/apiModule";
-import * as S from "./Question.style";
+import * as S from "./style/Question.style";
 
 const Question = ({ documentData }) => {
   const [questions, setQuestions] = useState([]);
