+++ src/pages/home/HomePage.jsx
@@ -1,19 +1,12 @@
 import React from "react";
-import styled from "styled-components";
 import * as S from "../../style/LayoutStyle";
 import Navbar from "../../components/common/Navbar";
-
-const VLine = styled.div`
-  border-left: 1px solid white;
-  min-height: 100vh;
-`;
 
 const HomePage = () => {
   return (
     <>
       <S.Layout>
         <Navbar />
-        <VLine></VLine>
         <S.HomeContainer>
           <img
             src={`${process.env.PUBLIC_URL}/LogoBlack.svg`}
