+++ src/pages/login/LoginPage.jsx
@@ -1,84 +1,8 @@
 import axios from "axios";
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import styled from "styled-components";
 import { setCookie } from "../../api/cookie";
-
-const LoginLayout = styled.div`
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-  justify-content: center;
-  align-items: center;
-
-  .logo {
-    width: 311px;
-    margin-bottom: 90px;
-  }
-`;
-
-const Form = styled.form`
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-  justify-content: center;
-  align-items: center;
-`;
-
-const InputWrapper = styled.div`
-  display: flex;
-  flex-direction: row;
-  width: 500px;
-  justify-content: space-between;
-  margin-bottom: 45px;
-`;
-
-const Div = styled.div`
-  color: #fff;
-  text-align: center;
-  font-size: 30px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-  letter-spacing: -1.515px;
-`;
-
-const StyledInput = styled.input`
-  width: 306px;
-  height: 45px;
-  border-radius: 15px;
-  border: 1px solid #fefefe;
-  background: rgba(254, 254, 254, 0);
-  color: #fff;
-  font-size: 24px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-  /* letter-spacing: -1.212px; */
-  outline: none;
-  padding-left: 10px;
-  &:focus {
-    outline: none;
-    border: 1px solid #eb9537;
-  }
-`;
-
-const LoginButton = styled.button`
-  margin-top: 70px;
-  width: 190px;
-  height: 57.881px;
-  flex-shrink: 0;
-  background: linear-gradient(180deg, #eb9537 5.52%, #ecc08f 96.15%);
-  border-radius: 10px;
-  border: none;
-  color: #000;
-  text-align: center;
-  font-size: 25px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-  letter-spacing: -1.263px;
-`;
+import * as S from "./style/LoginPage.style";
 
 const LoginPage = () => {
   const navigate = useNavigate();
@@ -123,37 +47,37 @@ const LoginPage = () => {
   };
 
   return (
-    <LoginLayout>
+    <S.LoginLayout>
       <img
         src={`${process.env.PUBLIC_URL}/Logo.svg`}
         alt="Logo"
         className="logo"
       />
-      <Form onSubmit={handleSubmit}>
-        <InputWrapper>
-          <Div>아이디</Div>
-          <StyledInput
+      <S.Form onSubmit={handleSubmit}>
+        <S.InputWrapper>
+          <S.Div>아이디</S.Div>
+          <S.StyledInput
             type="id"
             id="accountId"
             name="accountId"
             value={values.accountId}
             onChange={handleChange}
-          ></StyledInput>
-        </InputWrapper>
-        <InputWrapper>
-          <Div>비밀번호</Div>
-          <StyledInput
+          ></S.StyledInput>
+        </S.InputWrapper>
+        <S.InputWrapper>
+          <S.Div>비밀번호</S.Div>
+          <S.StyledInput
             type="password"
             id="password"
             name="password"
             value={values.password}
             onChange={handleChange}
             autoComplete="off"
-          ></StyledInput>
-        </InputWrapper>
-        <LoginButton type="submit">로그인</LoginButton>
-      </Form>
-    </LoginLayout>
+          ></S.StyledInput>
+        </S.InputWrapper>
+        <S.LoginButton type="submit">로그인</S.LoginButton>
+      </S.Form>
+    </S.LoginLayout>
   );
 };
