+++ src/pages/recruit/QuestionEditPage.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import * as L from "../../style/LayoutStyle";
 import * as B from "../../components/button/Button.style";
-import * as S from "./QuestionPage.style";
+import * as S from "./style/QuestionPage.style";
 import Navbar from "../../components/common/Navbar";
 import apiModule from "../../api/apiModule";
