+++ src/pages/recruit/style/ApplyStatusPage.style.jsx
