+++ src/pages/recruit/style/DocsDetailPage.style.jsx
